% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizBySubgroup-plotFcns.R
\name{collapseProp}
\alias{collapseProp}
\title{Calculates proportions of individuals in a 2x2 table of (trt, binary covar)}
\usage{
collapseProp(data, var = 1)
}
\arguments{
\item{data}{Any object that can be coerced into a \code{data.frame},
that contains the data needing to be summarized.
The first column is a binary covariate needing to be summarized.
The second column
is a binary variable named \code{trt}
that indicates each observation's treatment status (0 or 1).}

\item{var}{An scalar integer that indicates
which column(s) of \code{data} contain the covariates to be
summarized. \emph{This argument is deprecated and will
be removed in later versions.}}
}
\value{
A data frame with three variables:
  \itemize{
    \item \code{values}, contains the levels of \code{var}
    among the treated and the levels among the untreated
    \item \code{trt}, an factor variable, indicating
    which treatment group the
    row corresponds to
    \item \code{propor}, the proportion of the entire sample that has that
  }
}
\description{
\code{collapseProp()} is an internal function that takes a data
frame of a particular structure and returns a data frame with proportions
that are used in creating bar charts.
}
