% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFcns.R
\name{covarInDataset}
\alias{covarInDataset}
\title{Checking whether a specific covariate is present}
\usage{
covarInDataset(covarName, covarNames)
}
\arguments{
\item{covarName}{A scalar string, containing the name of the
covariate of interest.}

\item{covarNames}{A vector of strings, defining a list of
covariate names to check against.}
}
\value{
A logical scalar, indicating whether the covariate is
  contained within the list. Note that by definition,
  a \code{NULL} or empty string (\code{""}) will return
  \code{TRUE}.
}
\description{
This is a
simple internal function
designed to check whether a specific covariate is
contained within a list of covariates.
}
