% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarProfiles-plotFcn.R
\name{covarProfiles}
\alias{covarProfiles}
\title{Generates figure for 'Covariate Profiles' tab}
\usage{
covarProfiles(ds, CPplotPts = TRUE, CPplotLines = TRUE)
}
\arguments{
\item{ds}{Any object that can be coerced into a \code{data.frame},
that contains data needed for plotting.}

\item{CPplotPts}{A logical scalar. Should the plot include points?}

\item{CPplotLines}{A logical scalar. Should the plot include
connecting lines?}
}
\value{
If both of the logical arguments are false, the output
  will be a \strong{plotly} object containing an empty plot.
  Otherwise, the output will be a \strong{ggplot} object
  containing the plot.
}
\description{
\code{covarProfiles()} is an internal function that
takes a dataset and returns a plot.
}
\seealso{
Other plotting functions: \code{\link{forestPlotGen.fewGroups}},
  \code{\link{forestPlotGen.manyGroups}},
  \code{\link{forestPlot}}, \code{\link{groupQuantiles}},
  \code{\link{subgroupProfiles}}, \code{\link{vizByCovar}},
  \code{\link{vizBySubgroup}}
}
