% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFcns.R
\name{dataCompat}
\alias{dataCompat}
\title{Formatting data to be compatible wth plotting}
\usage{
dataCompat(ds, outcome, treatment, estITE = NULL, group)
}
\arguments{
\item{ds}{Data to be formatted.}

\item{outcome}{The column index of the outcome variable.}

\item{treatment}{The column index of the treatment variable.}

\item{estITE}{The column index of the estimated ITE variable.}

\item{group}{The column index of the grouping variable.}
}
\value{
The dataset with values in tact, but column names
        changed to be compatible with plotting functions
}
\description{
This is an
internal function
to rename covariates within the dataset.
}
