% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFcns.R
\name{expit}
\alias{expit}
\title{Expit scale transformation}
\usage{
expit(x)
}
\arguments{
\item{x}{A vector of real values.}
}
\value{
A vector of real values, the same length as the input.
}
\description{
Also known as the "inverse logit", this is an
internal function
to transform real line values to
logit scale values (i.e., \eqn{(0,1)}).
\eqn{expit(x)} is defined as \eqn{ e^x / (1 + e^x) }
for any real value.
}
