% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizBySubgroup-plotFcns.R
\name{flexPlot.byStratum}
\alias{flexPlot.byStratum}
\title{Dynamically produce the code for the "Viz by Subgroup" tab}
\usage{
flexPlot.byStratum(varNames = "", reset = FALSE)
}
\arguments{
\item{varNames}{A vector of plot names to be displayed.}

\item{reset}{A logical scalar. Should the tab be generated
with no constituent plots?}
}
\value{
A string such that, once evaluated, will generate the
  "Viz by Subgroup" tab.
}
\description{
\code{flexPlot.byStratum()} takes a vector of plot names
(as strings) and generates the "Viz by Subgroup"
tabPanel and placeholders for the constituent plots.
}
