% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestPlot-plotFcns.R
\name{forestPlot}
\alias{forestPlot}
\title{Generates figure for 'Forest Plot' tab}
\usage{
forestPlot(ds, displayMedian = TRUE, displayError = TRUE,
  displayLine = TRUE, displayMarMean = TRUE, displayBoxplot = TRUE,
  jitter = TRUE, col.outcome, col.trt, col.ITE, col.grp, simData,
  indivPlot = TRUE)
}
\arguments{
\item{ds}{Any object that can be coerced into a \code{data.frame},
that contains data needed for plotting. This dataset is
of a very specific structure, as
defined within \code{dataCompat()}.}

\item{displayMedian}{A logical scalar. In the case of subgroup-level
plotting (\code{indivPlot = FALSE}), should the median treatment
effect for
each subgroup be displayed?}

\item{displayError}{A logical scalar. In the case of subgroup-level
plotting (\code{indivPlot = FALSE}), should the error bars for
each subgroup be displayed?}

\item{displayLine}{A logical scalar. In the case of subgroup-level
plotting (\code{indivPlot = FALSE}), should a line connecting
the median from
each subgroup be displayed?}

\item{displayMarMean}{A logical scalar. In the case of subgroup-level
plotting (\code{indivPlot = FALSE}), should a vertical line denoting
the marginal mean be displayed?}

\item{displayBoxplot}{A logical scalar. In the case of individual-level
plotting (\code{indivPlot = TRUE}), should the boxplot for
each subgroup be displayed?}

\item{jitter}{A logical scalar. In the case of individual-level
plotting (\code{indivPlot = TRUE}), should the plot contain the jittered
individual-level data (typically estimated ITEs?)}

\item{col.outcome}{A scalar integer, indicating the column of the \code{ds}
that contains the outcome.}

\item{col.trt}{A scalar integer, indicating the column of the \code{ds}
that contains the (binary) treatment.}

\item{col.ITE}{A scalar integer, indicating the column of the \code{ds}
that contains the estimated ITEs.}

\item{col.grp}{A scalar integer, indicating the column of the \code{ds}
that contains the grouping indicator.}

\item{simData}{A logical scalar, indicating whether the data is the
simple simulated data
(\code{TRUE}) or is more complex (\code{FALSE}).}

\item{indivPlot}{A logical scalar, indicating whether to generate a plot
at the individual level (\code{TRUE}) or at the subgroup level
(\code{FALSE}).}
}
\value{
If both of the logical arguments are false, the output
  will be a \strong{plotly} object containing an empty plot.
  Otherwise, the output will be a \strong{ggplot} object
  containing the plot.
}
\description{
\code{forestPlot()} is an internal function that
takes a dataset and returns a plot.
}
\section{Details}{

  This function calls \code{shiny::incProgress()}, which is used
  to increment the progress bar within the UI.
}

\seealso{
Other plotting functions: \code{\link{covarProfiles}},
  \code{\link{forestPlotGen.fewGroups}},
  \code{\link{forestPlotGen.manyGroups}},
  \code{\link{groupQuantiles}},
  \code{\link{subgroupProfiles}}, \code{\link{vizByCovar}},
  \code{\link{vizBySubgroup}}
}
