% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestPlot-plotFcns.R
\name{forestPlotGen.fewGroups}
\alias{forestPlotGen.fewGroups}
\title{Generates individual-level figure for 'Forest Plot' tab}
\usage{
forestPlotGen.fewGroups(ds, displayBoxplot = TRUE, jitter = TRUE, simData)
}
\arguments{
\item{ds}{A dataset of a very specific structure, as
defined within \code{dataCompat()}}

\item{displayBoxplot}{A logical scalar. In the case of individual-level
plotting (\code{indivPlot = TRUE}), should the boxplot for
each subgroup be displayed?}

\item{jitter}{A logical scalar. In the case of individual-level
plotting (\code{indivPlot = TRUE}), should the plot contain the jittered
individual-level data (typically estimated ITEs?)}

\item{simData}{A logical scalar, indicating whether the data is the
simple simulated data}
}
\value{
If both of the logical arguments are false, the output
  will be a \strong{plotly} object containing an empty plot.
  Otherwise, the output will be a \strong{ggplot} object
  containing the plot.
}
\description{
\code{forestPlotGen.fewGroups()} is an internal function that
takes a dataset and returns a plot.
}
\section{Details}{

  This function calls \code{shiny::incProgress()}, which is used
  to increment the progress bar within the UI.
}

\seealso{
Other plotting functions: \code{\link{covarProfiles}},
  \code{\link{forestPlotGen.manyGroups}},
  \code{\link{forestPlot}}, \code{\link{groupQuantiles}},
  \code{\link{subgroupProfiles}}, \code{\link{vizByCovar}},
  \code{\link{vizBySubgroup}}
}
