% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestPlot-plotFcns.R
\name{forestPlotGen.manyGroups}
\alias{forestPlotGen.manyGroups}
\title{Generates individual-level figure for 'Forest Plot' tab}
\usage{
forestPlotGen.manyGroups(ds, displayMedian = TRUE, displayError = TRUE,
  displayLine = FALSE, displayMarMean = TRUE)
}
\arguments{
\item{ds}{A dataset of a very specific structure, as
defined within \code{dataCompat()}.}

\item{displayMedian}{A logical scalar. In the case of subgroup-level
plotting (\code{indivPlot = FALSE}), should the median treatment
effect for
each subgroup be displayed?}

\item{displayError}{A logical scalar. In the case of subgroup-level
plotting (\code{indivPlot = FALSE}), should the error bars for
each subgroup be displayed?}

\item{displayLine}{A logical scalar. In the case of subgroup-level
plotting (\code{indivPlot = FALSE}), should a line connecting
the median from
each subgroup be displayed?}

\item{displayMarMean}{A logical scalar. In the case of subgroup-level
plotting (\code{indivPlot = FALSE}), should a vertical line denoting
the marginal mean be displayed?}
}
\value{
If an untenable combination
  of the logical arguments are false, the output
  will be a \strong{plotly} object containing an empty plot.
  Otherwise, the output will be a \strong{ggplot} object
  containing the plot.
}
\description{
\code{forestPlotGen.manyGroups()} is an internal function that
takes a dataset and returns a plot.
}
\section{Details}{

  This function calls \code{shiny::incProgress()}, which is used
  to increment the progress bar within the UI.
}

\seealso{
Other plotting functions: \code{\link{covarProfiles}},
  \code{\link{forestPlotGen.fewGroups}},
  \code{\link{forestPlot}}, \code{\link{groupQuantiles}},
  \code{\link{subgroupProfiles}}, \code{\link{vizByCovar}},
  \code{\link{vizBySubgroup}}
}
