% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyApp.R
\name{hetviz}
\alias{hetviz}
\title{hetviz: Treatment effect heterogeneity (TEH) visualization}
\usage{
hetviz()
}
\description{
To launch the application, the user should call \code{hetviz()} (no
arguments) from the console. This call will launch
the eponymous \code{shiny} application locally,
within the default web browser of the user's machine.
There are no arguments in this call; all arguments will be provided to the package
through the graphical user interface.
}
