% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizByCovar-plotFcn.R
\name{vizByCovar}
\alias{vizByCovar}
\title{Generate figure for 'Viz by Covariate' tab}
\usage{
vizByCovar(ds, covar0, VCPplotPts = TRUE, VCPplotLines = FALSE)
}
\arguments{
\item{ds}{Any object that can be coerced into a \code{data.frame},
that contains data needed for plotting.}

\item{covar0}{A scalar integer, defining which column of the data set
contains the covariate to be plotted.}

\item{VCPplotPts}{A logical scalar. Should the plot include points?}

\item{VCPplotLines}{A logical scalar. Should the plot include
a connecting line?}
}
\value{
If both of the logical arguments are false, the output
  will be a \strong{plotly} object containing an empty plot.
  Otherwise, the output will be a \strong{ggplot} object
  containing the plot.
}
\description{
\code{vizByCovar()} is an internal function that generates a
plot containing the covariate profile of a specified
covariate.
}
\seealso{
Other plotting functions: \code{\link{covarProfiles}},
  \code{\link{forestPlotGen.fewGroups}},
  \code{\link{forestPlotGen.manyGroups}},
  \code{\link{forestPlot}}, \code{\link{groupQuantiles}},
  \code{\link{subgroupProfiles}},
  \code{\link{vizBySubgroup}}
}
