% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizBySubgroup-plotFcns.R
\name{vizBySubgroup}
\alias{vizBySubgroup}
\title{Generates figures for 'Viz by Subgroup' tab}
\usage{
vizBySubgroup(covarData, trt, covarName, ctgLevels = 5, plotHeight = 200,
  plotWidth = 300, trtColor = "#ffec5c", ctrlColor = "#e1315b")
}
\arguments{
\item{covarData}{A vector of numeric covariate data to plot.}

\item{trt}{A binary numeric vector that indicates each
observation's treatment status (0 or 1).}

\item{covarName}{A string that contains the name of the covariate
being plotted (for annotating the plot).}

\item{ctgLevels}{A scalar integer that defines the maximum number
of levels a covariate can have and still be considered
categorical.}

\item{plotHeight}{A scalar integer that defines the height of
the plot; passed to \code{ggvis()}.}

\item{plotWidth}{A scalar integer that defines the width of
the plot; passed to \code{ggvis()}.}

\item{trtColor}{The color of treated observations in the bar
chart. This argument is passed to \code{ggvis()} and can be
any color specification that's acceptable by R (e.g.,
see \href{https://www.stat.ubc.ca/~jenny/STAT545A/block14_colors.html}{here}
for more information).}

\item{ctrlColor}{The color of control observations in the bar
chart. This argument is passed to \code{ggvis()} and can be
any color specification that's acceptable by R (e.g.,
see \href{https://www.stat.ubc.ca/~jenny/STAT545A/block14_colors.html}{here}
for more information).}
}
\value{
If both of the logical arguments are false, the output
  will be a \strong{plotly} object containing an empty plot.
  Otherwise, the output will be a \strong{ggplot} object
  containing the plot.
}
\description{
\code{vizBySubgroup()} is an internal function that
takes a vector of covariate data, a vector of (binary) treatment
assignments, and generates a plot.
}
\seealso{
Other plotting functions: \code{\link{covarProfiles}},
  \code{\link{forestPlotGen.fewGroups}},
  \code{\link{forestPlotGen.manyGroups}},
  \code{\link{forestPlot}}, \code{\link{groupQuantiles}},
  \code{\link{subgroupProfiles}}, \code{\link{vizByCovar}}
}
