% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizByCovar-plotFcn.R
\name{vizByCovar}
\alias{vizByCovar}
\title{Generate figure for 'Viz by Covariate' tab}
\usage{
vizByCovar(ds, covar0, VCPplotPts = TRUE, VCPplotLines = FALSE, col.outcome,
  col.trt, col.ITE, col.grp, cols.covars, cols.cts, cols.ctg, simData)
}
\arguments{
\item{ds}{Any object that can be coerced into a \code{data.frame},
that contains data needed for plotting. This dataset is
of a very specific structure, as
defined within \code{dataCompat()}.}

\item{covar0}{A scalar integer, defining which column of the data set
contains the covariate to be plotted.}

\item{VCPplotPts}{A logical scalar. Should the plot include points?}

\item{VCPplotLines}{A logical scalar. Should the plot include
a connecting line?}

\item{col.outcome}{A scalar integer, indicating the column of \code{ds}
that contains the outcome.}

\item{col.trt}{A scalar integer, indicating the column of \code{ds}
that contains the (binary) treatment.}

\item{col.ITE}{A scalar integer, indicating the column of \code{ds}
that contains the estimated ITEs.}

\item{col.grp}{A scalar integer, indicating the column of \code{ds}
that contains the grouping indicator.}

\item{cols.covars}{A integer vector, indicating the columns of \code{ds}
that contain covariates (i.e., excluding treatment, outcome,
grouping var, and estimated ITE (if provided)).}

\item{cols.cts}{A integer vector, indicating the columns of \code{ds}
that contain continuous variables (including the treatment and outcome
variables, if applicable).}

\item{cols.ctg}{A integer vector, indicating the columns of \code{ds}
that contain polytomous variables (including the treatment and outcome
variables, if applicable).}

\item{simData}{A logical scalar, indicating whether the data is the
simple simulated data
(\code{TRUE}) or is more complex (\code{FALSE}).}
}
\value{
If both of the logical arguments are false, the output
  will be a \strong{plotly} object containing an empty plot.
  Otherwise, the output will be a \strong{ggplot} object
  containing the plot.
}
\description{
\code{vizByCovar()} is an internal function that generates a
plot containing the covariate profile of a specified
covariate.
}
\seealso{
Other plotting functions: \code{\link{covarProfiles}},
  \code{\link{forestPlotGen.fewGroups}},
  \code{\link{forestPlotGen.manyGroups}},
  \code{\link{forestPlot}}, \code{\link{groupQuantiles}},
  \code{\link{subgroupProfiles}},
  \code{\link{vizBySubgroup}}
}
